speedy/requirements.txt
django
gunicorn
django-cleanup
psycopg2-binary
django-modeltranslation
django-grappelli
pillow
django-robots
django-ckeditor